+++ src/app/components/share-show/edm.js
@@ -3,6 +3,7 @@
  *
  * @author Jakub Liput
  * @copyright (C) 2024 ACK CYFRONET AGH
+ * @copyright (C) 2026 Onedata (onedata.org)
  * @license This software is released under the MIT license cited in 'LICENSE.txt'.
  */
 
@@ -24,6 +25,7 @@ import EdmObjectType from '../../utils/edm/object-type';
 import EdmPropertyFactory from '../../utils/edm/property-factory';
 import { asyncObserver } from 'onedata-gui-common/utils/observer';
 import insufficientPrivilegesMessage from 'onedata-gui-common/utils/i18n/insufficient-privileges-message';
+import { guidFor } from '@ember/object/internals';
 
 const defaultMode = 'visual';
 
@@ -166,6 +168,12 @@ export default Component.extend(I18n, {
    * @type {boolean}
    */
   isSaving: false,
+
+  /**
+   * If set to true, user can submit the metadata without valid state.
+   * @type {boolean}
+   */
+  isValidationIgnored: false,
 
   //#endregion
 
@@ -340,6 +348,7 @@ export default Component.extend(I18n, {
     'isValid',
     'editMode',
     'visualEdmViewModel.isModified',
+    'isValidationIgnored',
     function submitDisabledReason() {
       if (this.modelXmlSyncState === EdmModelXmlSyncState.Waiting) {
         return this.t('submitDisabledReason.validatingSync');
@@ -350,13 +359,24 @@ export default Component.extend(I18n, {
       if (this.modelXmlSyncState !== EdmModelXmlSyncState.Synced) {
         return this.t('submitDisabledReason.xmlNotAccepted');
       }
-      if (!this.isValid) {
+      if (!this.isValid && !this.isValidationIgnored) {
         return this.t('submitDisabledReason.invalid');
       }
       if (this.editMode === 'edit' && !this.visualEdmViewModel.isModified) {
         return this.t('submitDisabledReason.noChanges');
       }
       return null;
+    }
+  ),
+
+  submitWarningIconTip: computed(
+    'submitDisabledReason',
+    'isValidationIgnored',
+    'isValid',
+    function submitWarningIconTip() {
+      if (!this.submitDisabledReason && !this.isValid && this.isValidationIgnored) {
+        return this.t('submitWarningIconReason.invalid');
+      }
     }
   ),
 
@@ -386,6 +406,10 @@ export default Component.extend(I18n, {
       return 'show';
     }
     return this.isPublished ? 'edit' : 'create';
+  }),
+
+  validationOptionsButtonId: computed(function validationOptionsButtonId() {
+    return guidFor(this) + '-validation-options-button';
   }),
 
   isEffDisabled: emberOr('isDisabled', 'isSaving'),
@@ -714,6 +738,14 @@ export default Component.extend(I18n, {
           this.visualEdmViewModel.markAsModified();
         }
       }
+    },
+
+    /**
+     * @param {boolean} value
+     * @returns {void}
+     */
+    changeValidationIgnored(value) {
+      this.set('isValidationIgnored', value);
     },
   },
 });